pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools>=61",
]

[project]
name = "replikant"
version = "0.1"
description = "A flexible toolkit to enable replicable subjective evaluations"
readme = {file="README.md", content-type = "text/markdown"}
authors = [
   {name="Sébastien Le Maguer", email="sebastien.lemaguer@helsinki.fi"}
]
requires-python = ">=3.11"
classifiers = [
  'Development Status :: 4 - Beta', # Audience
  'Intended Audience :: Science/Research', # Topics
  'License :: OSI Approved :: MIT License',
  "Programming Language :: Python :: 3 :: Only",
  'Programming Language :: Python :: 3.11',
  "Programming Language :: Python :: 3.12",
  'Topic :: Multimedia :: Sound/Audio :: Speech', # Pick your license as you wish
]
dependencies = [
  "cachelib>=0.7",
  "click>=8.1.3",
  "Flask>=2.1.2",
  "Flask-Session>=0.4",
  "Flask-SQLAlchemy>=2.5.1",
  "gevent",
  "greenlet>=1.1.2",
  "gunicorn",
  "importlib-metadata>=4.11.4",
  "itsdangerous>=2.1.2",
  "Jinja2>=3.1.2",
  "MarkupSafe>=2.1.1",
  "pandas>=2.2.1",
  "PyYAML>=6",
  "SQLAlchemy>=2.0.27",
  "Werkzeug>=2.1.2",
  "zipp>=3.8",
]
[project.optional-dependencies]
dev = [
  "pre-commit",
  "black",
  "basedpyright"
]
test = [
  "selenium",
]
[project.urls]
Homepage = "https://github.com/seblemaguer/replikant"
Issues = "https://github.com/seblemaguer/replikant/issues"
git = "https://github.com/seblemaguer/replikant.git"
[project.scripts]
replikant = "replikant.main:main"

[tool.setuptools.package-data]
replikant = ["**/assets/**/*.*", "**/templates/**/*.*"]  # FIXME: this is a bit hardcore here

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.ini
  |Dockerfile
  |Jenkinfile
)/
'''

[tool.flake8]
max-line-length = 120

[tool.basedpyright]
typeCheckingMode = "standard"
